Change schema from Presets.MEDIA to customized schema (json).
Because Textile does NOT support video mills for now, we have to create a ``/json` mill to save binary media content (encoded with Base64) and its metadata (proofs and signatures). You can check the new customized schema in ``/res/raw/mediant_block_schema.json` .

With this new customized schema, it is possible to save arbitrary media format to Textile threads. The content of the media file must be encoded with Base64 to string to save in the json file.

Warning: one possible failure might occur due to memory out of space. Because currently we encode the binary file by reading entire file to memory, which would be dangerous on mobile device.

Every Textile block only saves ONE json file in this app. Therefore, when retrieving file hash, the file index is always 0.